# docs/swagger.yaml
basePath: /book
definitions:
...
  request.OrderBrc20DoBidReq:
    properties:
      address:
        type: string
      amount:
        type: string
      inscriptionId:
        type: string
      inscriptionNumber:
        type: string
      net:
        description: livenet/signet/testnet
        type: string
      orderId:
        type: string
      psbtRaw:
        type: string
      tick:
        type: string
      value:
        type: integer
    type: object
  request.OrderBrc20PushReq:
    properties:
      address:
        type: string
      coinAmount:
        type: integer
      net:
        description: livenet/signet/testnet
        type: string
      orderState:
        allOf:
        - $ref: '#/definitions/model.OrderState'
        description: 1-create
      orderType:
        allOf:
        - $ref: '#/definitions/model.OrderType'
        description: 1-sell,2-buy
      psbtRaw:
        type: string
      tick:
        type: string
    type: object
...
  respond.BalanceDetails:
    properties:
      availableBalance:
        type: string
      balance:
        type: string
      limit:
        type: string
      page:
        type: string
      token:
        type: string
      tokenType:
        type: string
      totalPage:
        type: string
      transferBalance:
        type: string
      transferBalanceList:
        items:
          $ref: '#/definitions/respond.BalanceItem'
        type: array
    type: object
  respond.BalanceItem:
    properties:
      amount:
        type: string
      inscriptionId:
        type: string
      inscriptionNumber:
        type: string
    type: object
...
  respond.BidPsbt:
    properties:
      net:
        type: string
      orderId:
        type: string
      psbtRaw:
        description: PSBT Raw
        type: string
      tick:
        type: string
    type: object
  respond.Brc20BalanceList:
    properties:
      balanceList:
        items:
          $ref: '#/definitions/respond.BalanceListItem'
        type: array
      limit:
        type: string
      page:
        type: string
      totalPage:
        type: string
    type: object
...
  respond.Brc20Item:
    properties:
      amount:
        description: 'Btc: sat'
        type: integer
      buyerAddress:
        description: Buyer's address
        type: string
      coinAmount:
        description: Brc20 amount
        type: integer
      coinDecimalNum:
        description: omitempty
        type: integer
      coinRatePrice:
        description: Rate for brc20-btc
        type: integer
      decimalNum:
        description: Btc decimal
        type: integer
      freeState:
        allOf:
        - $ref: '#/definitions/model.FreeState'
        description: 1-for free
      inscriptionId:
        description: InscriptionId
        type: string
      net:
        description: Net env
        type: string
      orderId:
        description: Order ID
        type: string
      orderState:
        allOf:
        - $ref: '#/definitions/model.OrderState'
        description: Order state：1-create,2-finish,3-cancel
      orderType:
        allOf:
        - $ref: '#/definitions/model.OrderType'
        description: Order type：1-sell,2-buy
      psbtRaw:
        description: PSBT Raw
        type: string
      sellerAddress:
        description: Seller's address
        type: string
      tick:
        description: Brc20 symbol
        type: string
      timestamp:
        description: Create time
        type: integer
    type: object
...
  respond.CheckBrc20InscriptionReq:
    properties:
      actionType:
        type: string
      amount:
        type: string
      availableTransferState:
        description: 'Available Transfer state: success/fail'
        type: string
      blockHeight:
        type: string
      inscriptionId:
        type: string
      inscriptionNumber:
        type: string
      inscriptionState:
        description: 'inscribe state: success/fail'
        type: string
      location:
        description: location - txid:vout:offset
        type: string
      ownerAddress:
        type: string
      token:
        description: tick name
        type: string
      tokenType:
        description: token type
        type: string
      txId:
        type: string
    type: object
...
  respond.KlineItem:
    properties:
      close:
        type: string
      high:
        type: string
      low:
        type: string
      open:
        type: string
      timestamp:
        type: integer
      volume:
        type: integer
    type: object
  respond.Message:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      processingTime:
        type: integer
    type: object
...
info:
  contact:
    name: API Support
  description: OrdBook API Service
  title: OrdBook API Service
  version: "1.0"
paths:
  /brc20/address/{address}/{tick}:
    get:
      description: Check inscription brc20 valid
      parameters:
      - description: tick
        in: path
        name: tick
        required: true
        type: string
      - description: address
        in: path
        name: address
        required: true
        type: string
      - description: net
        in: query
        name: net
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BalanceDetails'
      summary: Check inscription brc20 valid
      tags:
      - brc20
  /brc20/address/{address}/balance/info:
    get:
      description: Get brc20 balance
      parameters:
      - description: tick
        in: query
        name: tick
        type: string
      - description: address
        in: path
        name: address
        required: true
        type: string
      - description: net
        in: query
        name: net
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Brc20BalanceList'
      summary: Get brc20 balance
      tags:
      - brc20
  /brc20/check/info:
    get:
      description: Check inscription brc20 valid
      parameters:
      - description: inscriptionId
        in: query
        name: inscriptionId
        type: string
      - description: inscriptionNumber
        in: query
        name: inscriptionNumber
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.CheckBrc20InscriptionReq'
      summary: Check inscription brc20 valid
      tags:
      - brc20
...
  /brc20/kline:
    get:
      description: Fetch tick kline data
      parameters:
      - description: net
        in: query
        name: net
        required: true
        type: string
      - description: tick
        in: query
        name: tick
        required: true
        type: string
      - description: interval：1m/1s/15m/1h/4h/1d/1w/
        in: query
        name: interval
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: flag
        in: query
        name: flag
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.KlineItem'
      summary: Fetch tick kline data
      tags:
      - brc20
  /brc20/order/{orderId}:
    get:
      description: Fetch one order
      parameters:
      - description: orderId
        in: path
        name: orderId
        required: true
        type: string
      - description: buyerAddress
        in: query
        name: buyerAddress
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Brc20Item'
      summary: Fetch one order
      tags:
      - brc20
  /brc20/order/ask/push:
    post:
      description: Push order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.OrderBrc20PushReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Push order
      tags:
      - brc20
  /brc20/order/bid:
    get:
      description: Fetch bid
      parameters:
      - description: net:mainnet/signet/testnet
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: inscriptionId
        in: query
        name: inscriptionId
        type: string
      - description: inscriptionNumber
        in: query
        name: inscriptionNumber
        type: string
      - description: coinAmount
        in: query
        name: coinAmount
        type: string
      - description: address
        in: query
        name: address
        type: string
      - description: amount
        in: query
        name: amount
        type: integer
      - description: isPool for pool
        in: query
        name: isPool
        type: boolean
      - description: orderId of pool
        in: query
        name: orderId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.BidPsbt'
      summary: Fetch bid
      tags:
      - brc20
  /brc20/order/bid/do:
    post:
      description: Do bid order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.OrderBrc20DoBidReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Do bid order
      tags:
      - brc20
  /brc20/order/bid/dummy/{address}:
    get:
      description: Get bid dummy list
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      - description: net
        in: query
        name: net
        type: string
      - description: skip
        in: query
        name: skip
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Brc20BidDummyResponse'
      summary: Get bid dummy list
      tags:
      - brc20
  /brc20/order/bid/pre:
    get:
      description: Fetch pre bid
      parameters:
      - description: net:mainnet/signet/testnet
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: 'limit: Max-50'
        in: query
        name: limit
        type: integer
      - description: flag
        in: query
        name: flag
        type: integer
      - description: isPool
        in: query
        name: isPool
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderResponse'
      summary: Fetch pre bid
      tags:
      - brc20
  /brc20/order/bid/push:
    post:
      description: Push bid order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.OrderBrc20UpdateBidReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Push bid order
      tags:
      - brc20
  /brc20/order/update:
    post:
      description: Update order
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/request.OrderBrc20UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Message'
      summary: Update order
      tags:
      - brc20
  /brc20/orders:
    get:
      description: Fetch orders
      parameters:
      - description: net:mainnet/signet/testnet
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: sellerAddress
        in: query
        name: sellerAddress
        type: string
      - description: buyerAddress
        in: query
        name: buyerAddress
        type: string
      - description: 'orderState: 1-create,2-finish,3-cancel,5-timeout,6-err,100-all'
        in: query
        name: orderState
        type: integer
      - description: 'orderType: 1-sell,2-buy'
        in: query
        name: orderType
        type: integer
      - description: 'limit: Max-50'
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: flag
        in: query
        name: flag
        type: integer
      - description: 'sortKey: timestamp/coinRatePrice, default:timestamp'
        in: query
        name: sortKey
        type: string
      - description: 'sortType: 1/-1'
        in: query
        name: sortType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderResponse'
      summary: Fetch orders
      tags:
      - brc20
  /brc20/orders/user/{address}:
    get:
      description: Fetch user orders
      parameters:
      - description: net:mainnet/signet/testnet
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      - description: address
        in: path
        name: address
        required: true
        type: string
      - description: 'orderState: 1-create,2-finish,3-cancel,5-timeout,6-err,100-all'
        in: query
        name: orderState
        type: integer
      - description: 'orderType: 1-sell,2-buy'
        in: query
        name: orderType
        type: integer
      - description: 'limit: Max-50'
        in: query
        name: limit
        type: integer
      - description: flag
        in: query
        name: flag
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: 'sortKey: timestamp/coinRatePrice, default:timestamp'
        in: query
        name: sortKey
        type: string
      - description: 'sortType: 1/-1'
        in: query
        name: sortType
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.OrderResponse'
      summary: Fetch user orders
      tags:
      - brc20
...
  /brc20/tickers:
    get:
      description: Fetch tick info
      parameters:
      - description: net
        in: query
        name: net
        type: string
      - description: tick
        in: query
        name: tick
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respond.Brc20TickInfoResponse'
      summary: Fetch tick info
      tags:
      - brc20
...
schemes:
- https
swagger: "2.0"